# docs/FE-contract/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HistoryTalk API",
    "description": "API Specification for HistoryTalk project, synced with FE and latest DB specifications.",
    "version": "2.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Development Server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/characters/context/{contextId}": {
      "get": {
        "tags": ["Characters"],
        "summary": "Lấy danh sách nhân vật thuộc bối cảnh lịch sử. CUSTOMER: chỉ nhận published. ADMIN/STAFF: nhận tất cả",
        "parameters": [
          { "name": "contextId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseWrapper" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Character" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/characters/{characterId}/contexts/{contextId}": {
      "post": {
        "tags": ["Characters"],
        "summary": "Gắn nhân vật vào bối cảnh lịch sử",
        "parameters": [
          { "name": "characterId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "contextId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Thành công"
          }
        }
      }
    },
    "/historical-contexts": {
      "get": {
        "tags": ["Historical Contexts"],
        "summary": "Lấy danh sách bối cảnh lịch sử (Phân trang)",
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Tìm theo tên" },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 6 } },
          { "name": "era", "in": "query", "schema": { "$ref": "#/components/schemas/Era" } }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseWrapper" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "allOf": [
                            { "$ref": "#/components/schemas/Pagination" },
                            {
                              "type": "object",
                              "properties": {
                                "content": {
                                  "type": "array",
                                  "items": { "$ref": "#/components/schemas/HistoricalContext" }
                                }
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Historical Contexts"],
        "summary": "Tạo bối cảnh lịch sử mới (CONTENT_ADMIN | SYSTEM_ADMIN)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoricalContextUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tạo thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseWrapper" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/HistoricalContext" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/historical-contexts/{id}": {
      "get": {
        "tags": ["Historical Contexts"],
        "summary": "Chi tiết bối cảnh lịch sử",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseWrapper" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/HistoricalContext" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Historical Contexts"],
        "summary": "Cập nhật bối cảnh lịch sử",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoricalContextUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseWrapper" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/HistoricalContext" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Historical Contexts"],
        "summary": "Xóa vĩnh viễn bối cảnh",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Thành công"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Era": {
        "type": "string",
        "enum": ["ANCIENT", "MEDIEVAL", "MODERN", "CONTEMPORARY"]
      },
      "ResponseWrapper": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" },
          "data": { "type": "object" },
          "timestamp": { "type": "string", "format": "date-time" }
        },
        "required": ["success", "message", "timestamp"]
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "totalElements": { "type": "integer" },
          "totalPages": { "type": "integer" },
          "currentPage": { "type": "integer" },
          "pageSize": { "type": "integer" },
          "hasNext": { "type": "boolean" },
          "hasPrevious": { "type": "boolean" }
        },
        "required": ["totalElements", "totalPages", "currentPage", "pageSize", "hasNext", "hasPrevious"]
      },
...
      "Character": {
        "type": "object",
        "properties": {
          "characterId": { "type": "string" },
          "name": { "type": "string" },
          "title": { "type": "string" },
          "background": { "type": "string" },
          "image": { "type": "string", "nullable": true },
          "personality": { "type": "string" },
          "lifespan": { "type": "string" },
          "era": { "type": "string" },
          "isActive": { "type": "boolean" },
          "isPublished": { "type": "boolean", "description": "true = đã publish cho người dùng xem" },
          "createdAt": { "type": "string", "format": "date-time", "description": "Thời gian tạo - chỉ trả cho ADMIN/STAFF" },
          "updatedAt": { "type": "string", "format": "date-time", "description": "Thời gian cập nhật - chỉ trả cho ADMIN/STAFF" },
          "context": {
            "type": "object",
            "properties": {
              "contextId": { "type": "string" }
            }
          }
        },
        "required": ["characterId", "name", "title", "background", "isActive", "isPublished"]
      },
...
      "HistoricalContext": {
        "type": "object",
        "properties": {
          "contextId": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "year": { "type": "integer" },
          "yearLabel": { "type": "string" },
          "era": { "type": "string" },
          "location": { "type": "string" },
          "imageUrl": { "type": "string", "nullable": true },
          "videoUrl": { "type": "string", "nullable": true },
          "period": { "type": "string" },
          "isActive": { "type": "boolean" }
        },
        "required": ["contextId", "name", "description", "year", "era", "isActive"]
      },
      "HistoricalContextUpsertRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "era": { "$ref": "#/components/schemas/Era" },
          "year": { "type": "integer" },
          "location": { "type": "string" },
          "imageUrl": { "type": "string", "nullable": true },
          "videoUrl": { "type": "string", "nullable": true },
          "isActive": { "type": "boolean" }
        },
        "required": ["name", "description", "era", "year", "isActive"]
      },
...
    }
  }
}